Fix: Pydantic Validation Error in data usuage over 24hr and Daily Byte Icon change

+++ backend/app/models/schemas.py
@@ -46,7 +46,7 @@ class UnifiStatus(BaseStatus):
     devices_online: int = 0
     devices_offline: int = 0
     wireless_clients: int = 0
-    data_usage_24h: int = 0
+    data_usage_24h: float = 0.0
     wan_latency: float = 0.0
